requirements.txt
fastapi
uvicorn
pydantic
passlib
PyJWT
requests
python-multipart
jinja2